requirements.txt
chevron
boolean.py
lxml
pyyaml
-e rand_soc
